requirements.txt
requests
python-dotenv
plotly

